Fix incorrect suffix search for alternation with RegexOptions.RightToLeft (#101408)
* Fix incorrect suffix search for alternation with RegexOptions.RightToLeft

Our limited RegexOptions.RightToLeft support for prefix optimizations is mishandling alternations. For RTL, this actually needs to create a suffix, but for alternations, we were actually creating a prefix. As this option is rarely used and it's not worth significant investment in optimizing, the fix is just to disable the handling of alternations for RTL, as the support has been broken since it was introduced in .NET 7.

* Fix new tests

* Avoid annoying skip message