When initializing Lock in NativeAOT, get the processor count directly instead of using Environment (#101514)
It's a bit safer to get the proc count directly since the Environment properties involve class construction. We used to get the proc count directly before, and there's still a reason to do so to get the correct value with the current initialization pattern due to in-thread reentrancy of the same class construction.