JIT: Enable cross-block local assertion prop for natural loops
For natural loop headers we can allow propagating assertions from the
entering predecessors without intersecting them with the assertions from
the backedges, provided that we take care to kill all assertions about
locals that may be defined within the loop. We can pay a bit of TP to
compute this set before morph.

This is a similar to what VN does for natural loops, just less general.