Fix check for AppleClang in src/native/libs and fix warnings
While looking at https://github.com/dotnet/runtime/pull/101781 I noticed that we weren't catching AppleClang in the CMAKE_C_COMPILER_ID in src/native/libs/CMakeLists.txt because we used  STREQUAL instead of MATCHES like we do everywhere else.

The result was that we didn't enable `-Weverything` on Apple platforms for the libs.native subset and a bunch of warnings didn't get enabled, this PR fixes that.