Fix handling exceptions on shutdown
A recent partial backport of a change to enable propagating exceptions
on shutdown was missing two places in the code that were under a
different condition than the global exception handling disabling that
was removed. I must have made a mistake when testing that change (not
setting watson registry settings properly or using a stale build).
This change makes it work correctly.