Fix arm64 callee saved register set
When I've made a change to optimize the collided exception handling by
copying the stack frame iterator state, I've somehow missed adding two
callee saved registers, x27 and x28 to a list of callee saved registers
and that caused them to not to be copied during the state copying.

It has caused failures in one of the coreclr tests when running with
specific JIT stress settings.

Close #100476