[aot] handle case of partial generic sharing for unsafe accessor
In partial generic sharing, a call to an instance like `Foo<int>` is
replaced by `Foo<T_INT>` where T is constrained to `int` and enums
that use `int` as a base type.

In that case the AOT compiler will emit the unsafe accessor wrapper
instantiated with `T_INT`.  So in the AOT lookup we have to find calls
to `UnsafeAccessor<int>` and replace them with calls to `(wrapper)
UnsafeAccessor<T_INT>` to match what the AOT compiler is doing