JIT: Allow using pinvoke registers in the prolog
This code seems to be a leftover from a time where we generated pinvoke
helper init calls during the prolog. Today we create these calls as
normal IR in scratch BBs and merged return BBs, which means that it
should be just fine to use these registers during the prolog.

This frees up a bunch of registers to be used for the init register.
Previously we could be in a situation where we had no register available
for the init register on x86 when EAX was taken for the "special stub
arg".

Fix #102161