Modify error message for failed RW mappings
The executable allocator can fail to create a RW mapping for an existing
RX mapping due to both being out of memory and exceeding maximum number
of memory mappings. People were getting confused by the original error
message.
This chnage updates it to explicitly mention the cases when it can
occur.