Report catch type dependencies in IL scanner
We were missing reporting the type used in a catch when it doesn't require a generic lookup. This can be a problem when such type is never thrown and the exception is generic (in which case we sometimes upgrade necessary types to constructed types and those need scanned vtables).

Found in the Pri 1 test suite.