Add ability to dump ILLink traces for all tests (#102409)
This extends the `--dump-dependencies` option to take an optional
argument that specifies an assembly name. If none is specified,
the behavior is unchanged (trace will include everything),
otherwise the trace will only include members from the mentioned
assembly.

This also adds the ability to run ILLink tests in two new modes:

- `/p:GenerateExpectedDependencyTraces=true` will dump
  dependencies for all testcases to a location alongside the test
  source files (in a Dependencies subfolder).

- `/p:ValidateDependencyTraces=true` will look for the expected
  dependency traces in the source locations, dump dependencies
  for all testcases into the test output directory, and validate
  them against the expected dependency traces.

I couldn't think of a great way to automate this validation
without checking in a bunch of XML files or introducing more
infrastructure to save the traces corresponding to a commit, so
for now the intended use is for local validation.