Allow non-IMemberDefinition sources to be origin for event methods (#102613)
There are some cases where an event may be marked by something that is not an IMemberDefinition, for example, an assembly level attribute. https://github.com/dotnet/runtime/pull/102528 made an assumption that this was not possible and threw in those cases. This reverts the change related to that assumption.

Co-authored-by: Sven Boemer <sbomer@gmail.com>