Stop guessing if a type is blittable (#102645)
Fixes #75666.

CoreLib was making an assumption that if a type didn't have pregenerated interop data and it didn't have GC pointers, it was probably blittable. This PR stops making that guess. We're generating an AOT warning whenever the compiler is not sure we have interop data to make an API call succeed. It should be fine to just say "I don't have interop data" instead of trying too hard to make the API call succeed.